test(integration): share live agent lifecycle assertions
